fixed some jwtService bugs

+++ src/main/java/com/example/trellite/repository/AuthRepo.java
@@ -3,6 +3,6 @@ package com.example.trellite.repository;
 import com.example.trellite.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuthRepo extends JpaRepository<User, Integer> {
+public interface AuthRepo extends JpaRepository<User, Long> {
     User findByUsername(String username);
 }
